Low-Level Functions

ProgressiveHedging.add_leaf

ProgressiveHedging.add_node

ProgressiveHedging.apply_to_subproblem

ProgressiveHedging.branch_value

ProgressiveHedging.cb

ProgressiveHedging.consensus_variables

ProgressiveHedging.convert_to_variable_ids

ProgressiveHedging.convert_to_xhat_id

ProgressiveHedging.get_callback

ProgressiveHedging.get_callback_ext

ProgressiveHedging.index

ProgressiveHedging.is_integer

ProgressiveHedging.is_leaf

ProgressiveHedging.leaf_value

ProgressiveHedging.lower_bounds

ProgressiveHedging.mean_deviation

ProgressiveHedging.name

ProgressiveHedging.print_timing

ProgressiveHedging.probability

ProgressiveHedging.residuals

ProgressiveHedging.retrieve_aug_obj_value

ProgressiveHedging.retrieve_no_hat_history

ProgressiveHedging.retrieve_no_hats

ProgressiveHedging.retrieve_obj_value

ProgressiveHedging.retrieve_soln

ProgressiveHedging.retrieve_w

ProgressiveHedging.retrieve_w_history

ProgressiveHedging.retrieve_xhat_history

ProgressiveHedging.root

ProgressiveHedging.scenario

ProgressiveHedging.scenario_bundle

ProgressiveHedging.scenario_tree

ProgressiveHedging.scenarios

ProgressiveHedging.scid

ProgressiveHedging.spcb

ProgressiveHedging.stage

ProgressiveHedging.stage_id

ProgressiveHedging.stid

ProgressiveHedging.two_stage_tree

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.value

ProgressiveHedging.variable_fixing

ProgressiveHedging.variables

ProgressiveHedging.w_value

ProgressiveHedging.xhat_value

Callback Creation and Retrieval

ProgressiveHedging.apply_to_subproblem — Function

apply_to_subproblem(to_apply::Function,
                    phd::PHData,
                    winf::WorkerInf,
                    scid::ScenarioID,
                    args::Tuple=(),
                    kwargs::NamedTuple=NamedTuple(),
                    )

Applies the function to_apply to the subproblem with scenario id scid.

ProgressiveHedging.cb — Function

cb(f::Function)
cb(f::Function, ext::Dict{Symbol,Any})
cb(f::Function, initialize::Function)
cb(f::Function, initialize::Function, ext::Dict{Symbol,Any})
cb(name::String, f::Function, ext::Dict{Symbol,Any})

Shorthand for Callback functions with the same signature.

ProgressiveHedging.get_callback — Function

get_callback(phd::PHData, name::String)::Callback

Retrieve the callback with name name.

ProgressiveHedging.get_callback_ext — Function

get_callback_ext(phd::PHData, name::String)::Dict{Symbol,Any}

Retrieve the external dictionary for callback name.

ProgressiveHedging.mean_deviation — Function

mean_deviation(;tol::Float64=1e-8,
               save_deviations::Bool=false
               )::Callback

Implementation of a termination criterion given in section 2.3 of (Watson & Woodruff 2010). There it is called 'td'.

ProgressiveHedging.spcb — Function

spcb(f::Function)
spcb(f::Function, ext::Dict{Symbol,Any})
spcb(name::String, f::Function, ext::Dict{Symbol,Any})

Shorthand for SubproblemCallback.

ProgressiveHedging.variable_fixing — Function

variable_fixing(; lag=2, eq_tol=1e-8)::Callback

Implementation of the variable fixing convergence acceleration heuristic in section 2.2 of (Watson & Woodruff 2010).

ID Type Interactions

ProgressiveHedging.convert_to_variable_ids — Function

convert_to_variable_ids(phd::PHData, xid::XhatID)::Set{VariableID}

Convert the given consensus variable id to the contributing individual subproblem variable ids.

Arguments

phd::PHData : PH data structure for the corresponding problem

xid::XhatID : consensus variable id to convert

ProgressiveHedging.convert_to_xhat_id — Function

convert_to_xhat_id(phd::PHData, vid::VariableID)::XhatID

Convert the given VariableID to the consensus variable id (XhatID).

Arguments

phd::PHData : PH data structure for the corresponding problem

vid::VariableID : variable id to convert

ProgressiveHedging.index — Function

index(vid::VariableID)::Index

Returns the Index of the specified VariableID.

index(n::Int64)::Index

Create Index from n.

ProgressiveHedging.scenario — Function

scenario(vid::VariableID)::ScenarioID

Returns the ScenarioID of the specified VariableID.

ProgressiveHedging.scid — Function

scid(n::Int64)::ScenarioID

Create ScenarioID from n.

ProgressiveHedging.stage — Function

stage(vid::VariableID)::StageID

Returns the StageID of the specified VariableID.

ProgressiveHedging.stage_id — Function

stage_id(phd::PHData, xid::XhatID)::StageID

Returns the StageID in which the given consensus variable is.

ProgressiveHedging.stid — Function

stid(n::Int64)::StageID

Create StageID from n.

ProgressiveHedging.value — Method

value(idx::Index)

Return the raw (non-type safe) value of an Index

ProgressiveHedging.value — Method

value(nid::NodeID)

Return the raw (non-type safe) value of a NodeID

ProgressiveHedging.value — Method

value(scid::ScenarioID)

Return the raw (non-type safe) value of a ScenarioID

ProgressiveHedging.value — Method

value(stid::StageID)

Return the raw (non-type safe) value of a StageID

Result Retrieval

ProgressiveHedging.lower_bounds — Function

lower_bounds(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the computed and saved lower bounds. See the lower_bound keyword argument on solve.

ProgressiveHedging.print_timing — Function

print_timing([io::IO], phd::PHData)

Prints timining information from the Progressive Hedging solve.

ProgressiveHedging.residuals — Function

residuals(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with absolute and relative residuals along with the components. See the save_residuals keyword argument on solve.

ProgressiveHedging.retrieve_soln — Function

retrieve_soln(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the solution values of all consensus and leaf variables.

ProgressiveHedging.retrieve_aug_obj_value — Function

retrieve_aug_obj_value(phd::PHData)::Float64

Return the current objective value including Lagrange and proximal PH terms of the stochastic program.

ProgressiveHedging.retrieve_obj_value — Function

retrieve_obj_value(phd::PHData)::Float64

Return the current objective value without Lagrange and proximal PH terms.

ProgressiveHedging.retrieve_no_hats — Function

retrieve_no_hats(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the final nonconsensus variable values.

ProgressiveHedging.retrieve_no_hat_history — Function

retrieve_no_hat_history(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the saved nonconsensus variable values. See the save_iterates keyword argument on solve.

ProgressiveHedging.retrieve_w — Function

retrieve_w(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the final anticipativity constraint Lagrange multiplier values.

ProgressiveHedging.retrieve_w_history — Function

retrieve_w_history(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the saved PH Lagrange variable values. See the save_iterates keyword argument on solve.

ProgressiveHedging.retrieve_xhat_history — Function

retrieve_xhat_history(phd::PHData)::DataFrames.DataFrame

Return a DataFrame with the saved consensus variable values. See the save_iterates keyword argument on solve.

Problem Accessors

ProgressiveHedging.probability — Function

probability(phd::PHData, scenario::ScenarioID)::Float64

Returns the probability of the given scenario.

ProgressiveHedging.scenario_tree — Function

scenario_tree(phd::PHData)::ScenarioTree

Returns the scenario tree for created stochastic programming problem.

ProgressiveHedging.scenarios — Function

scenarios(phd::PHData)::Set{ScenarioID}

Returns the set of all scenarios for the stochastic problem.

Scenario Tree

ProgressiveHedging.add_node — Function

add_node(tree::ScenarioTree, parent::ScenarioNode)

Add a node to the ScenarioTree tree with parent node parent. Return the added node. If the node to add is a leaf, use add_leaf instead.

ProgressiveHedging.add_leaf — Function

add_leaf(tree::ScenarioTree, parent::ScenarioNode, probability<:Real)

Add a leaf to the ScenarioTree tree with parent node parent. The probability of this scenario occuring is given by probability. Returns the ScenarioID representing the scenario.

ProgressiveHedging.root — Function

root(tree::ScenarioTree)

Return the root node of the given ScenarioTree

ProgressiveHedging.two_stage_tree — Function

two_stage_tree(n::Int)::ScenarioTree
two_stage_tree(p::Vector{R})::ScenarioTree where R <: Real

Construct and return a two-stage scenario tree with n scenarios with equal probability or a two-stage scenario tree with length(p) scenarios where probability of scenario s is p[s+1].

Variable Interactions

ProgressiveHedging.branch_value — Function

branch_value(phd::PHData, vid::VariableID)::Float64

Returns the value of the variable associated with vid. Must be a branch variable.

See also: leaf_value, value

branch_value(phd::PHData, scen::ScenarioID, stage::StageID, idx::Index)::Float64

Returns the value of the variable associated with with scenario scen, stage stage and index idx. Must be a branch variable.

See also: leaf_value, value

ProgressiveHedging.consensus_variables — Function

consensus_variables(phd::PHData)::Dict{XhatID,HatVariable}

Returns the collection of consensus variables for the problem.

ProgressiveHedging.is_integer — Function

is_integer(a::HatVariable)::Bool

Returns true if the consensus variable is an integer variable. The consensus variable is an integer if the contributing subproblem variables are all integer variables.

ProgressiveHedging.is_leaf — Function

is_leaf(phd::PHData, xhid::XhatID)::Bool

Returns true if the given consensus variable id belongs to a leaf vertex in the scenario tree.

ProgressiveHedging.leaf_value — Function

leaf_value(phd::PHData, vid::VariableID)::Float64
leaf_value(phd::PHData, scen::ScenarioID, stage::StageID, idx::Index)::Float64

Returns the value of the variable associated with vid or with scenario scen, stage stage and index idx. Must be a leaf variable.

WARNING: For computational efficiency, leaf values are collected only at the end of a PH run. Therefore, using this function in a callback will result in an error.

See also: branch_value, value

ProgressiveHedging.name — Function

name(phd::PHData, vid::VariableID)::String

Returns the name of the consensus variable for the given VariableID.

name(phd::PHData, xid::XhatID)::String

Returns the name of the consensus variable for the given XhatID. The name is the same given to the individual scenario variables.

ProgressiveHedging.scenario_bundle — Function

scenario_bundle(phd::PHData, xid::XhatID)::Set{ScenarioID}

Returns the scenarios contributing to the consensus variable associated with xid.

ProgressiveHedging.value — Method

value(a::HatVariable)::Float64

Returns the current value of a.

ProgressiveHedging.value — Method

value(phd::PHData, vid::VariableID)

Returns the value of the variable associated with vid.

See also: branch_value, leaf_value

ProgressiveHedging.value — Method

value(phd::PHData, scen::ScenarioID, stage::StageID, idx::Index)::Float64

Returns the value of the variable associated with scenario scen, stage stage and index idx.

See also: branch_value, leaf_value

ProgressiveHedging.variables — Function

variables(a::HatVariable)::Set{VariableID}

Returns the variable ids for all subproblem variables contributing to this variable.

ProgressiveHedging.w_value — Function

w_value(phd::PHData, vid::VariableID)

Returns the value of the variable associated with vid. Only available for branch variables.

w_value(phd::PHData, scen::ScenarioID, stage::StageID, idx::Index)::Float64

Returns the value of the variable associated with scenario scen, stage stage and index idx. Only available for branch variables.

ProgressiveHedging.xhat_value — Function

xhat_value(phd::PHData, xhid::VariableID)::Float64

Returns the value of the consensus variable associated with xhid. Only available for leaf variables after calling solve. Available for branch variables at any time.

xhat_value(phd::PHData, vid::VariableID)::Float64

Returns the value of the consensus variable associated with vid. Only available for leaf variables after calling solve. Available for branch variables at any time.

xhat_value(phd::PHData, scen::ScenarioID, stage::StageID, idx::Index)::Float64

Returns the value of the consensus variable associated with scenario scen, stage stage and index idx. Only available for leaf variables after calling solve. Available for branch variables at any time.